requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_py-xgboost-mutex=2.0=cpu_0
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
altair=4.1.0=py_1
arviz=0.7.0=py_0
asn1crypto=1.3.0=py37_0
astor=0.8.0=py37_0
atari-py=0.2.6=pypi_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
binutils_impl_linux-64=2.34=h53a641e_0
binutils_linux-64=2.34=hc952b39_18
blas=1.0=openblas
bleach=3.1.0=pypi_0
blinker=1.4=py37_0
box2d-py=2.3.8=pypi_0
bzip2=1.0.8=h516909a_2
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.1.1=0
cachetools=3.1.1=py_0
cairo=1.14.12=h8948797_3
catboost=0.22=py37_0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37h2e261b9_0
cftime=1.1.1.2=py37h03ebfcd_0
chardet=3.0.4=py37_1003
click=7.1.1=py_0
cloudpickle=1.2.2=pypi_0
convertdate=2.1.3=py_1000
cryptography=2.8=py37h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
curl=7.69.1=h33f0ec9_0
cvxpy=1.0.28=pypi_0
cycler=0.10.0=py_2
cython=0.29.15=pypi_0
cytoolz=0.10.1=py37h516909a_0
dask-core=2.14.0=py_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=pypi_0
defusedxml=0.6.0=pypi_0
dill=0.3.1.1=pypi_0
ecos=2.0.7.post1=pypi_0
eli5=0.10.1=py37_1
entrypoints=0.3=pypi_0
ephem=3.7.7.1=py37h516909a_0
expat=2.2.6=he6710b0_0
fbprophet=0.6=py37he1b5a44_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.1=he06d7ca_0
fribidi=1.0.9=h516909a_0
future=0.18.2=pypi_0
gast=0.2.2=py37_0
gcc_impl_linux-64=7.3.0=habb00fd_1
gcc_linux-64=7.3.0=h553295d_18
glfw=1.11.0=pypi_0
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
google-auth=1.13.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
graphite2=1.3.13=he1b5a44_1001
graphviz=2.40.1=h21bd128_2
grpcio=1.27.2=py37hf8bcb03_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gxx_impl_linux-64=7.3.0=hdf63c60_1
gxx_linux-64=7.3.0=h553295d_18
gym=0.16.0=dev_0
h5py=2.10.0=py37h7918eee_0
harfbuzz=1.8.8=hffaf4a1_0
hdf4=4.2.13=hf30be14_1003
hdf5=1.10.4=hb1b8bf9_0
holidays=0.10.1=py_0
icu=58.2=hf484d3e_1000
idna=2.8=pypi_0
imageio=2.8.0=py_0
importlib_metadata=1.3.0=py37_0
ipykernel=5.1.3=pypi_0
ipython=7.11.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.0=pypi_0
jedi=0.15.2=pypi_0
jinja2=2.10.3=pypi_0
joblib=0.14.1=py_0
jpeg=9c=h14c3975_1001
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter-client=5.3.4=pypi_0
jupyter-console=6.0.0=pypi_0
jupyter-core=4.6.1=pypi_0
jupyter_client=6.1.2=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.3=py37_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37h99015e2_0
krb5=1.17.1=h173b8e3_0
ld_impl_linux-64=2.34=h53a641e_0
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libcurl=7.69.1=hf7181ac_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_5
liblapack=3.8.0=14_openblas
libllvm8=8.0.1=hc9558a2_0
libnetcdf=4.7.3=hb80b6cc_0
libopenblas=0.3.7=h5ec1e0e_6
libpng=1.6.37=hed695b0_1
libprotobuf=3.11.4=hd408876_0
libsodium=1.0.16=h1bed415_0
libssh2=1.8.2=h22169c7_2
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=hc7e4089_6
libuuid=1.0.3=h1bed415_2
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h1bed415_1
libxgboost=1.0.2=he1b5a44_0
libxml2=2.9.9=hea5a465_1
lightgbm=2.3.0=py37he6710b0_0
llvmlite=0.31.0=py37h5202443_1
lockfile=0.12.2=pypi_0
lunarcalendar=0.0.9=py_0
lz4-c=1.8.3=he1b5a44_1001
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.2=pypi_0
matplotlib-base=3.1.3=py37hef1b27d_0
mistune=0.8.4=py37h7b6447c_0
more-itertools=8.0.2=py_0
multiprocess=0.70.9=pypi_0
nbconvert=5.6.1=py37_0
nbformat=5.0.3=pypi_0
ncurses=6.1=he6710b0_1
netcdf4=1.5.3=py37hbf33ddf_0
networkx=2.4=py_1
notebook=6.0.2=pypi_0
numba=0.48.0=py37hb3f55d8_0
numpy=1.18.0=pypi_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
open3d-python=0.7.0.0=pypi_0
opencv-python=4.2.0.32=pypi_0
openssl=1.1.1g=h7b6447c_0
opt_einsum=3.1.0=py_0
osqp=0.6.1=pypi_0
packaging=20.0=py_0
pandas=0.25.3=pypi_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
pango=1.42.4=h049681c_0
parso=0.5.2=pypi_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=pypi_0
pgmpy=0.1.9=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.0.0=py37hb39fc2d_0
pip=19.3.1=py37_0
pixman=0.38.0=h516909a_1003
plotly=4.6.0=py_0
pluggy=0.13.1=py37_0
prometheus_client=0.7.1=py_0
prompt-toolkit=2.0.10=pypi_0
prompt_toolkit=3.0.4=0
protobuf=3.11.4=py37he6710b0_0
ptyprocess=0.6.0=pypi_0
py=1.8.1=py_0
py-xgboost=1.0.2=py37hc8dfbb8_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pyglet=1.5.0=pypi_0
pygments=2.5.2=pypi_0
pyjwt=1.7.1=py37_0
pymdptoolbox=4.0b3=pypi_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.15.7=pypi_0
pysocks=1.7.1=py37_0
pystan=2.19.1.1=py37hb3f55d8_1
pytest=5.3.2=py37_0
python=3.7.5=h0371630_0
python-dateutil=2.8.1=py_0
python-graphviz=0.13.2=py_0
python_abi=3.7=1_cp37m
pytz=2019.3=py_0
pywavelets=1.1.1=py37h03ebfcd_1
pyzmq=18.1.1=pypi_0
qt=5.9.7=h5867ecd_1
qtconsole=4.6.0=pypi_0
qtpy=1.9.0=py_0
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=py_0
retrying=1.3.3=py37_2
rsa=4.0=py_0
scikit-image=0.16.2=py37hb3f55d8_0
scikit-learn=0.22.1=py37h22eb022_0
scipy=1.4.1=py37ha3d9a3c_3
scs=2.1.1-2=pypi_0
seaborn=0.10.0=py_0
send2trash=1.5.0=pypi_0
setuptools=44.0.0=py37_0
shap=0.35.0=py37h0da4684_0
singledispatch=3.4.0.3=py37_1000
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
sqlite=3.30.1=h7b6447c_0
tabulate=0.8.7=pyh9f0ad1d_0
tensorboard=2.1.0=py3_0
tensorflow=2.1.0=gpu_py37h7a4bb67_0
tensorflow-base=2.1.0=gpu_py37h6c5654b_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-gpu=2.1.0=h0d30ee6_0
termcolor=1.1.0=py37_1
terminado=0.8.3=pypi_0
testpath=0.4.4=py_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torch=1.4.0=pypi_0
tornado=6.0.3=pypi_0
tqdm=4.43.0=pypi_0
traitlets=4.3.3=py37_0
urllib3=1.25.7=pypi_0
vega=3.2.0=py37hc8dfbb8_0
vega_datasets=0.8.0=py_0
wcwidth=0.1.8=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=py_0
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.12.1=py37h7b6447c_1
xarray=0.15.1=py_0
xgboost=1.0.2=py37h3340039_0
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h3b9ef0a_2
